+++ advancedSettings.pro
@@ -59,7 +59,7 @@ SOURCES += src/main.cpp\
     src/tabcontrollers/audiomanager/AudioManagerDummy.cpp \
     src/tabcontrollers/keyboardinput/KeyboardInputDummy.cpp \
     src/overlaycontroller/openvr_init.cpp \
-    src/ivrinput/steamivrinput.cpp
+    src/ivrinput/ivrinput.cpp
 
 HEADERS += src/overlaycontroller.h \
     src/tabcontrollers/AudioTabController.h \
@@ -83,7 +83,7 @@ HEADERS += src/overlaycontroller.h \
     src/ivrinput/ivrinput_action.h \
     src/ivrinput/ivrinput_manifest.h \
     src/ivrinput/ivrinput_action_set.h \
-    src/ivrinput/steamivrinput.h
+    src/ivrinput/ivrinput.h
 
 win32 {
     SOURCES += src/tabcontrollers/audiomanager/AudioManagerWindows.cpp \
